Accurate time on devices that lack RTC
This update fixes inaccurate times on devices that lack an RTC. It does this by calculating a time offset between the local system and the server, as well as allowing a timezone to be configured in config. This basically allows the app to be deployed to any device without ever configuring it's clock and it will render the proper time.

The state variable `date` was moved up a layer because it was not in a place that was easy to update. It is passed into Header via props. A new prop was added to the header for `use24Hour` format. This creates a pathway in for future settings.